refactor: 💡 revert to plural, update config to align
