+++ design-toolkit/components/src/components/query-builder/__fixtures__/example-configuration.ts
